yb-load-panda/requirements2.txt
greenlet==1.0.0
importlib-metadata==3.7.3
numpy==1.20.1
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.3
typing-extensions==3.7.4.3
zipp==3.4.1
